+++ MetaRay/src/Hittable.h
@@ -9,7 +9,7 @@
 struct hit_record {
     vec3 p;
     vec3 normal;
-    double t;
+    float t;
     bool front_face;
 
     inline void set_face_normal(const ray& r, const vec3& outward_normal) {
@@ -20,7 +20,7 @@ struct hit_record {
 
 class Hittable {
 public:
-    virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
+    virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
 };
 
 class HitList : public Hittable {
@@ -30,7 +30,7 @@ public:
     inline void add(std::shared_ptr<Hittable> object) { m_Objects.push_back(object); }
     inline void clear() { m_Objects.clear(); }
 
-    virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const override;
+    virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const override;
 
 public:
     std::vector<std::shared_ptr<Hittable>> m_Objects;
