+++ MetaRay/src/Shapes.h
@@ -4,10 +4,10 @@
 
 class Sphere : public Hittable {
 public:
-	Sphere(const vec3& center, double radius);
-	bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const;
+	Sphere(const vec3& center, float radius);
+	bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const;
 
 private:
 	vec3 m_Center;
-	double m_Radius;
+	float m_Radius;
 };
